New Website Being Developed
This was a project that I worked on back in the spring 2023 semester. It was for my capstone project for American History. This was one of the more involved web development projects I've worked on, and this was still before I knew a lick of PHP...all the code was written by hand.

It's fun to see how far I've gone.